@@ -2,29 +2,57 @@
   "name": "@solx/source",
   "version": "0.0.0",
   "license": "MIT",
-  "scripts": {
-    "contracts:build": "nx build contracts"
-  },
   "private": true,
+  "scripts": {
+    "serve:api": "nx serve api",
+    "serve:contracts": "nx serve contracts",
+    "serve:indexer": "nx serve indexer",
+    "serve:web": "nx serve web",
+    "build:api": "nx build api",
+    "build:contracts": "nx build contracts",
+    "build:indexer": "nx build indexer",
+    "build:web": "nx build web",
+    "lint:api": "nx lint api",
+    "lint:api:fix": "nx lint api --fix",
+    "lint:contracts": "nx lint contracts",
+    "lint:contracts:fix": "nx lint contracts --fix",
+    "lint:indexer": "nx lint indexer",
+    "lint:indexer:fix": "nx lint indexer --fix",
+    "lint:web": "nx lint web",
+    "lint:web:fix": "nx lint web --fix"
+  },
   "dependencies": {
     "@coral-xyz/anchor": "^0.29.0",
+    "@nestjs/axios": "^4.0.0",
     "@nestjs/common": "^10.0.2",
+    "@nestjs/config": "^4.0.2",
     "@nestjs/core": "^10.0.2",
     "@nestjs/platform-express": "^10.0.2",
+    "@nestjs/terminus": "^11.0.0",
+    "@nestjs/throttler": "^6.4.0",
+    "@prisma/client": "^6.6.0",
     "axios": "^1.6.0",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.1",
+    "compression": "^1.8.0",
+    "cookie-parser": "^1.4.7",
+    "helmet": "^8.1.0",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-router-dom": "6.29.0",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.8.0"
+    "rxjs": "^7.8.0",
+    "zod": "^3.24.3"
   },
   "devDependencies": {
     "@coral-xyz/anchor": "^0.29.0",
-    "@eslint/js": "^9.8.0",
+    "@eslint/js": "^9.25.1",
     "@nestjs/schematics": "^10.0.1",
+    "@nestjs/swagger": "^11.1.5",
     "@nestjs/testing": "^10.0.2",
     "@nx/eslint": "20.8.0",
-    "@nx/eslint-plugin": "20.8.0",
+    "@nx/eslint-plugin": "^20.8.0",
+    "@nx/jest": "20.8.0",
     "@nx/js": "20.8.0",
     "@nx/nest": "20.8.0",
     "@nx/node": "20.8.0",
@@ -37,8 +65,13 @@
     "@swc/cli": "~0.6.0",
     "@swc/core": "~1.5.7",
     "@swc/helpers": "~0.5.11",
+    "@swc/jest": "~0.2.36",
+    "@tanstack/eslint-plugin-query": "^5.73.3",
     "@types/bn.js": "^5.1.0",
     "@types/chai": "^4.3.0",
+    "@types/compression": "^1.7.5",
+    "@types/cookie-parser": "^1.4.8",
+    "@types/jest": "^29.5.14",
     "@types/mocha": "^9.0.0",
     "@types/node": "^20.0.0",
     "@types/react": "19.0.0",
@@ -47,26 +80,37 @@
     "@vitest/ui": "^3.0.0",
     "autoprefixer": "10.4.13",
     "chai": "^4.3.4",
-    "eslint": "^9.8.0",
-    "eslint-config-prettier": "^10.0.0",
-    "eslint-plugin-import": "2.31.0",
+    "eslint": "^9.25.1",
+    "eslint-config-prettier": "^10.1.2",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jsx-a11y": "6.10.1",
-    "eslint-plugin-react": "7.35.0",
-    "eslint-plugin-react-hooks": "5.0.0",
+    "eslint-plugin-prettier": "^5.2.6",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-svg-jsx": "^1.2.4",
+    "jest": "^29.7.0",
+    "jest-environment-node": "^29.7.0",
     "jiti": "2.4.2",
     "jsdom": "~22.1.0",
+    "jsonc-eslint-parser": "^2.1.0",
     "mocha": "^9.0.3",
     "nx": "20.8.0",
     "postcss": "8.4.38",
-    "prettier": "^2.6.2",
+    "prettier": "^3.5.3",
+    "prisma": "^6.6.0",
     "tailwindcss": "3.4.3",
+    "ts-jest": "^29.1.0",
     "ts-mocha": "^10.0.0",
+    "ts-node": "10.9.1",
     "tslib": "^2.3.0",
     "typescript": "~5.7.2",
-    "typescript-eslint": "^8.19.0",
+    "typescript-eslint": "^8.31.0",
     "vite": "^6.0.0",
     "vitest": "^3.0.0",
     "webpack-cli": "^5.1.4"
+  },
+  "prisma": {
+    "schema": "./libs/data-access/prisma/schema"
   },
   "packageManager": "yarn@4.9.1"
 }
