@@ -1,5 +1,6 @@
 const { NxAppWebpackPlugin } = require('@nx/webpack/app-plugin');
-const { join } = require('path');
+
+const { join } = require('node:path');
 
 module.exports = {
   output: {
