@@ -1,6 +1,6 @@
-import NxWelcome from './nx-welcome';
-
 import { Route, Routes, Link } from 'react-router-dom';
+
+import NxWelcome from './nx-welcome';
 
 export function App() {
   return (
